Moved user retrieval to a custom middleware
