+++ app.js
@@ -18,6 +18,7 @@ app.use(cors());
 app.use(express.json());
 app.use(middleware.morganConfig);
 app.use(middleware.tokenExtractor);
+app.use(middleware.userExtractor);
 
 app.use("/api/blogs", blogsRouter);
 app.use("/api/users", usersRouter);
