+++ utils/middleware.js
@@ -1,6 +1,9 @@
+const jwt = require("jsonwebtoken");
 const morgan = require("morgan");
 
+const errors = require("./errors");
 const logger = require("./logger");
+const User = require("../models/user");
 
 const requestBodyToken = morgan.token(
   // eslint-disable-next-line no-unused-vars
@@ -15,6 +18,31 @@ const tokenExtractor = (request, response, next) => {
   const authorization = request.get("authorization");
   if (authorization && authorization.startsWith("Bearer ")) {
     request["token"] = authorization.replace("Bearer ", "");
+  }
+  next();
+};
+
+const userExtractor = async (request, response, next) => {
+  if (request.token) {
+    try {
+      const decodedToken = jwt.verify(
+        request.token, process.env.JWT_SECRET_SIGNATURE
+      );
+
+      if (!decodedToken.id) {
+        throw new errors.InvalidTokenError(decodedToken);
+      }
+
+      const user = await User.findById(decodedToken.id);
+
+      if (!user) {
+        throw new errors.InvalidUserError(user);
+      }
+
+      request["user"] = user;
+    } catch (error) {
+      next(error);
+    }
   }
   next();
 };
@@ -56,6 +84,7 @@ module.exports = {
   errorHandler,
   morganConfig,
   requestBodyToken,
+  tokenExtractor,
   unknownEndpoint,
-  tokenExtractor
+  userExtractor
 };
